services/crawler/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
boto3==1.34.0
apscheduler==3.10.4
python-dotenv==1.0.1
feedparser==6.0.11
